+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/ReactionsRow.kt
@@ -21,7 +21,6 @@
 package com.vitorpamplona.amethyst.ui.note
 
 import android.content.Context
-import android.content.Intent
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedContentTransitionScope
 import androidx.compose.animation.AnimatedVisibility
@@ -148,6 +147,7 @@ import com.vitorpamplona.amethyst.ui.components.toasts.multiline.UserBasedErrorM
 import com.vitorpamplona.amethyst.ui.navigation.navs.INav
 import com.vitorpamplona.amethyst.ui.navigation.routes.Route
 import com.vitorpamplona.amethyst.ui.navigation.routes.routeReplyTo
+import com.vitorpamplona.amethyst.ui.note.elements.ShareOptionsBottomSheet
 import com.vitorpamplona.amethyst.ui.note.types.EditState
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.header.PaymentTargetsDialog
@@ -318,6 +318,9 @@ private fun InnerReactionRow(
                     if (!isPrivateRumor) {
                         ShareReaction(
                             note = baseNote,
+                            editState = editState,
+                            accountViewModel = accountViewModel,
+                            nav = nav,
                             grayTint = MaterialTheme.colorScheme.placeholderText,
                         )
                     }
@@ -338,37 +341,32 @@ private fun InnerReactionRow(
 @Composable
 fun ShareReaction(
     note: Note,
+    editState: State<GenericLoadable<EditState>>?,
+    accountViewModel: AccountViewModel,
+    nav: INav,
     grayTint: Color,
     barChartModifier: Modifier = Size19Modifier,
 ) {
-    val context = LocalContext.current
+    var showShareSheet by remember { mutableStateOf(false) }
 
     ClickableBox(
         modifier = barChartModifier,
-        onClick = {
-            val sendIntent =
-                Intent().apply {
-                    action = Intent.ACTION_SEND
-                    type = "text/plain"
-                    putExtra(
-                        Intent.EXTRA_TEXT,
-                        externalLinkForNote(note),
-                    )
-                    putExtra(
-                        Intent.EXTRA_TITLE,
-                        stringRes(context, R.string.quick_action_share_browser_link),
-                    )
-                }
-
-            val shareIntent =
-                Intent.createChooser(
-                    sendIntent,
-                    stringRes(context, R.string.quick_action_share),
-                )
-            context.startActivity(shareIntent)
-        },
+        onClick = { showShareSheet = true },
     ) {
         ShareIcon(barChartModifier, grayTint)
+    }
+
+    if (showShareSheet) {
+        ShareOptionsBottomSheet(
+            note = note,
+            // The reaction-row Share button is only rendered for non-private
+            // notes, so the image/link share rows are always offered here.
+            isPrivateRumor = false,
+            editState = editState,
+            accountViewModel = accountViewModel,
+            nav = nav,
+            onDismiss = { showShareSheet = false },
+        )
     }
 }
