+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/elements/DropDownMenu.kt
@@ -20,17 +20,13 @@
  */
 package com.vitorpamplona.amethyst.ui.note.elements
 
-import android.content.Intent
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.platform.LocalClipboard
-import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.vitorpamplona.amethyst.R
 import com.vitorpamplona.amethyst.commons.icons.symbols.MaterialSymbols
@@ -42,19 +38,16 @@ import com.vitorpamplona.amethyst.ui.components.GenericLoadable
 import com.vitorpamplona.amethyst.ui.components.M3ActionDialog
 import com.vitorpamplona.amethyst.ui.components.M3ActionRow
 import com.vitorpamplona.amethyst.ui.components.M3ActionSection
-import com.vitorpamplona.amethyst.ui.components.util.setText
 import com.vitorpamplona.amethyst.ui.navigation.navs.INav
 import com.vitorpamplona.amethyst.ui.navigation.routes.Route
 import com.vitorpamplona.amethyst.ui.navigation.routes.routeEditDraftTo
 import com.vitorpamplona.amethyst.ui.note.VerticalDotsIcon
-import com.vitorpamplona.amethyst.ui.note.externalLinkForNote
 import com.vitorpamplona.amethyst.ui.note.types.EditState
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.report.ReportNoteDialog
 import com.vitorpamplona.amethyst.ui.stringRes
 import com.vitorpamplona.amethyst.ui.theme.Size24Modifier
 import com.vitorpamplona.quartz.experimental.music.track.MusicTrackEvent
-import com.vitorpamplona.quartz.nip01Core.jackson.JacksonMapper
 import com.vitorpamplona.quartz.nip01Core.tags.aTag.isTaggedAddressableNote
 import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
 import com.vitorpamplona.quartz.nip23LongContent.LongTextNoteEvent
@@ -64,8 +57,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 
 @Composable
 fun MoreOptionsButton(
@@ -157,10 +148,6 @@ fun NoteDropDownMenu(
         title = stringRes(R.string.note_actions_dialog_title),
         onDismiss = onDismiss,
     ) {
-        val clipboardManager = LocalClipboard.current
-        val actContext = LocalContext.current
-        val scope = rememberCoroutineScope()
-
         // Unsealed rumors (private replies/posts received in gift wraps) are
         // unsigned and must never be referenced by a public event: hide every
         // action that would publish an e-tag of this note (broadcast, edit,
@@ -191,65 +178,14 @@ fun NoteDropDownMenu(
 
         // Copy & Share section
         M3ActionSection {
-            M3ActionRow(icon = MaterialSymbols.ContentCopy, text = stringRes(R.string.copy_text)) {
-                val lastNoteVersion = (editState?.value as? GenericLoadable.Loaded)?.loaded?.modificationToShow?.value ?: note
-                accountViewModel.decrypt(lastNoteVersion) {
-                    scope.launch {
-                        clipboardManager.setText(it)
-                    }
-                }
-                onDismiss()
-            }
-            M3ActionRow(icon = MaterialSymbols.ContentCopy, text = stringRes(R.string.copy_user_pubkey)) {
-                note.author?.let {
-                    scope.launch(Dispatchers.IO) {
-                        clipboardManager.setText("nostr:${it.pubkeyNpub()}")
-                        onDismiss()
-                    }
-                }
-            }
-            M3ActionRow(icon = MaterialSymbols.ContentCopy, text = stringRes(R.string.copy_note_id)) {
-                scope.launch(Dispatchers.IO) {
-                    clipboardManager.setText(note.toNostrUri())
-                    onDismiss()
-                }
-            }
-            M3ActionRow(icon = MaterialSymbols.ContentCopy, text = stringRes(R.string.copy_raw_json)) {
-                val event = note.event
-                if (event != null) {
-                    scope.launch {
-                        val json = withContext(Dispatchers.Default) { JacksonMapper.toJsonPretty(event) }
-                        clipboardManager.setText(json)
-                        onDismiss()
-                    }
-                } else {
-                    onDismiss()
-                }
-            }
-            if (!isPrivateRumor) {
-                M3ActionRow(icon = MaterialSymbols.Share, text = stringRes(R.string.quick_action_share)) {
-                    val sendIntent =
-                        Intent().apply {
-                            action = Intent.ACTION_SEND
-                            type = "text/plain"
-                            putExtra(Intent.EXTRA_TEXT, externalLinkForNote(note))
-                            putExtra(Intent.EXTRA_TITLE, stringRes(actContext, R.string.quick_action_share_browser_link))
-                        }
-                    val shareIntent = Intent.createChooser(sendIntent, stringRes(actContext, R.string.quick_action_share))
-                    actContext.startActivity(shareIntent)
-                    onDismiss()
-                }
-                M3ActionRow(icon = MaterialSymbols.Image, text = stringRes(R.string.share_as_image)) {
-                    val shareId = if (note is AddressableNote) note.address.toValue() else note.idHex
-                    nav.nav(Route.ShareNoteAsImageFile(shareId))
-                    onDismiss()
-                }
-                M3ActionRow(icon = MaterialSymbols.Image, text = stringRes(R.string.share_as_image_url)) {
-                    val shareId = if (note is AddressableNote) note.address.toValue() else note.idHex
-                    nav.nav(Route.ShareNoteAsImage(shareId))
-                    onDismiss()
-                }
-            }
+            ShareCopyActionRows(
+                note = note,
+                isPrivateRumor = isPrivateRumor,
+                editState = editState,
+                accountViewModel = accountViewModel,
+                nav = nav,
+                onDismiss = onDismiss,
+            )
         }
 
         // Edit & Broadcast section
